authorizing delete

+++ App/controllers/ListingController.php
@@ -5,6 +5,8 @@ namespace App\Controllers;
 use Error;
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController{
     protected $db;
@@ -23,7 +25,7 @@ class ListingController{
 
     public function index(){
 
-        $listings = $this->db->query("SELECT * FROM listings")->fetchAll();
+        $listings = $this->db->query("SELECT * FROM listings ORDER BY created_at DESC")->fetchAll();
 
         loadView('listings/index', [
             'listings' => $listings
@@ -86,7 +88,7 @@ class ListingController{
         ];
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 2;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -153,11 +155,19 @@ class ListingController{
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        // Check if listing exists
         if(!$listing){
             ErrorController::notFound('Listing not found.');
             return;
         }
+
+        // Authorization
+        if(!Authorization::isOwner($listing->user_id)){
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing.';
+            return redirect('/listings/' . $listing->id);
+        }
         
+
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
         // Set flash message
